Remove Apple integration concern
